MM-67320 Move WithTooltip into shared package without modification
