+++ webapp/platform/shared/src/components/tooltip/tooltip_shortcut.test.tsx
