+++ webapp/platform/shared/src/components/tooltip/with_tooltip.scss
